backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Anthropic
anthropic>=0.36.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# Vector DB
chromadb>=0.6.0

# AI / embeddings
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
sentence-transformers>=3.3.0

# Redis
redis>=5.2.0
hiredis>=3.1.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.20

# Observability
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-otlp>=1.29.0
opentelemetry-instrumentation-fastapi>=0.50b0
prometheus-fastapi-instrumentator>=7.0.0

# Utilities
httpx>=0.28.0
python-dotenv>=1.0.0
structlog>=24.4.0

# Dev / test
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
httpx>=0.28.0
ruff>=0.9.0
